add new users to db

+++ bot/database/database.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, Date
+from sqlalchemy import Column, String, Integer, Date, BIGINT
 from sqlalchemy.ext.declarative import declarative_base
 
 
@@ -7,7 +7,7 @@ Base = declarative_base()
 
 class User(Base):
     __tablename__ = "User"
-    tgid = Column(Integer, primary_key=True, index=True)
+    tgid = Column(BIGINT, primary_key=True, index=True)
     username = Column(String(length=255))
     need_notif = Column(Integer, default=1)
 
@@ -16,7 +16,7 @@ class Table(Base):
     __tablename__ = "Table"
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    tgid = Column(Integer)
+    tgid = Column(BIGINT)
     day_of_week = Column(String(length=15))
     date = Column(Date)
     timestart = Column(String(length=5), nullable=True)
